Add a few more tiny scales. Allow user to input doubles in preferences.

+++ ApplicationLibCode/Application/RiaGuiApplication.cpp
@@ -1403,8 +1403,8 @@ void RiaGuiApplication::applyGuiPreferences( const RiaPreferences*              
                         existingViewsWithCustomColors = true;
                     }
                     if ( m_preferences->defaultScaleFactorZ() != oldPreferences->defaultScaleFactorZ() &&
-                         rim3dView->scaleZ() != static_cast<double>( oldPreferences->defaultScaleFactorZ() ) &&
-                         rim3dView->scaleZ() != static_cast<double>( m_preferences->defaultScaleFactorZ() ) )
+                         rim3dView->scaleZ() != oldPreferences->defaultScaleFactorZ() &&
+                         rim3dView->scaleZ() != m_preferences->defaultScaleFactorZ() )
                     {
                         existingViewsWithCustomZScale = true;
                     }
@@ -1471,10 +1471,9 @@ void RiaGuiApplication::applyGuiPreferences( const RiaPreferences*              
                     rim3dView->applyBackgroundColorAndFontChanges();
                 }
 
-                if ( oldPreferences &&
-                     ( applySettingsToAllViews || rim3dView->scaleZ() == static_cast<double>( oldPreferences->defaultScaleFactorZ() ) ) )
+                if ( oldPreferences && ( applySettingsToAllViews || rim3dView->scaleZ() == oldPreferences->defaultScaleFactorZ() ) )
                 {
-                    rim3dView->setScaleZ( static_cast<double>( m_preferences->defaultScaleFactorZ() ) );
+                    rim3dView->setScaleZ( m_preferences->defaultScaleFactorZ() );
                     rim3dView->updateScaling();
                     if ( rim3dView == activeViewWindow() )
                     {
